Makefile: Create single image for imx6 socs based boards
IMX6 platform has two stage boot loaders like SPL and
U-Boot proper. For each stage we need to burn the image
on to flash with respective offsets.

This patch create a single image using binman, so that
user can get rid of burning different stage boot images.

without this patch:
------------------
$ sudo dd if=SPL of=/dev/mmcblk0 bs=1k seek=1
$ sudo dd if=u-boot-dtb.img of=/dev/mmcblk0 bs=1k seek=69

with this patch:
---------------
$ sudo dd if=u-boot-imx6-with-spl.bin of=/dev/mmcblk0 bs=1k seek=1

This would be easily extended to single image creation
for other imx6 soc boards.

This was tested on engicam imx6qdl and imx6ul boards

Signed-off-by: Shyam Saini <shyam.saini@amarulasolutions.com>